python/requirements.txt
sequana-fastqc
biopython